# task2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/repos/{owner}/{repo}": {
            "get": {
                "description": "Возвращает базовую информацию о публичном репозитории GitHub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repos"
                ],
                "summary": "Получить информацию о репозитории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Владелец репозитория (username или организация)",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название репозитория",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о репозитории",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Репозиторий не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "Превышен лимит запросов к GitHub",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}